% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_YEARMODA.R
\name{add_YEARMODA}
\alias{add_YEARMODA}
\title{Add a column for YEARMODA (or YEARMODAHO) to a data frame}
\usage{
add_YEARMODA(data, add_hour = FALSE)
}
\arguments{
\item{data}{is a data frame at least containing the columns \code{Year}, \code{Month}, and \code{Day} 
in the case of daily records; and \code{Year}, \code{Month}, \code{Day}, and \code{Hour} in the case
of hourly records. For now, column names must be spelled as shown in lines above}

\item{add_hour}{is a boolean parameter defining whether the output should add the \code{YEARMODAHO} column.
Default is set to \code{FALSE} so, even in case of hourly records, the ouput data frame will only show the 
\code{YEARMODA} column}
}
\value{
A data frame
}
\description{
This function allows users to add a column called YEARMODA or YEARMODAHO to any \code{\link{chillR}}-formated 
data frame
}
\examples{

df <- data.frame(Year = rep(2011, 10),
                 Month = rep(1, 10),
                 Day = seq(1, 10),
                 Tmin = c(10, 12, 11, NA, 21, NA, NA, 10, 20, 23))
                 
add_YEARMODA(data = df, add_hour = FALSE)

}
