% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chile_weather.R
\name{chile_weather}
\alias{chile_weather}
\title{Get historical records of extreme temperatures for Chile}
\usage{
chile_weather(output, Initial_Date = "1950-01-01",
  End_Date = "2017-12-31", latitude = latitude,
  longitude = longitude, Number_of_stations = 25,
  path_zip_tmin = path_zip_tmin, path_zip_tmax = path_zip_tmax)
}
\arguments{
\item{output}{Character string input which define the kind of data wanted. There are three options
for this parameter. "info_stations" provides a dataframe with information about a given number
of weather stations (set in "Number_of_stations" parameter) located close to the location
established with "latitude" and "longitude" parameters. "station_list_data" provides a list of
dataframes containing minimum and maximum temperature records from each weather station
obtained with the "info_stations" option within the period established in the call of the function.
Finally, "my_data" provides the first dataframe of the previous list and represent the
data of the closest weather station to the ubication established.}

\item{Initial_Date}{Character string input in the form "YYYY-MM-DD". This parameter represents the
initial date of the period of interest. If it is not provided, the default is established as
"1950-01-01" which correspond to the oldest date posible to use.}

\item{End_Date}{Character string input in the form "YYYY-MM-DD". This parameter represents the
final date of the period of interest. If it is not provided, the default is established as
"2017-12-31" which corresponds to the earliest posible date that can be used for the assesment.}

\item{latitude}{Numerical input. Latitude of the site of interest in decimal format.}

\item{longitude}{Numerical input. Longitude of the site of interest in decimal format.}

\item{Number_of_stations}{Numerical input. Number of stations wanted to use as closest stations to
the site of interest. Default option is the value 25.}

\item{path_zip_tmin}{Character string input. Location of the zip file containing minimum
temperatures. This input must include the name and extension of the file.}

\item{path_zip_tmax}{Character string input. Location of the zip file containing maximum
temperatures. This input must include the name and extension of the file.}
}
\description{
This function allows to obtain information about weather stations located in Chile as well as
a data frame containing historical records of minimum and maximum temperatures from those weather
stations. This function works with data downloaded from the website of the Center for Climate and
Resilience Research (CR)2 sponsored by the University of Chile. The function works for previously
downloaded data in ".zip" format. Data can be downloaded from the following links:
<http://www.cr2.cl/download/cr2_tasmindaily_2018_ghcn-zip/?wpdmdl=15125> for minimum temperatures and
<http://www.cr2.cl/download/cr2_tasmaxdaily_2018_ghcn-zip/?wpdmdl=15126> for maximum temperatures.
Function require both zip files in the same folder.
}
\examples{
#Getting the location of zip files
#path_zip_tmin<-"[Your folder]\\\\cr2_tasminDaily_2018_ghcn.zip"
#path_zip_tmax<-"[Your folder]\\\\cr2_tasmaxDaily_2018_ghcn.zip"

#Call of the function
#chile_weather(output = "my_data", Initial_Date = "2000-01-01", End_Date = "2017-12-31",
#                latitude = -32.8958, longitude = -71.2092, Number_of_stations = 25,
#                path_zip_tmin = path_zip_tmin, path_zip_tmax = path_zip_tmax)


}
