% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chill_days.R
\name{chill_days}
\alias{chill_days}
\title{Chill days (Cesaraccio et al. 2004)}
\usage{
chill_days(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing colums "Tmax" and "Tmin". These values must correspond to daily
records.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
This function computes the Chill Days according to the model proposed by Cesaraccio et al. (2004) developed
for several tree species including pears, kiwifruit and sweet cherries. In this function chill days are
computed without the negative (-) tranformation made in the original paper.
}
\references{
Cesaraccio C., Spano D., Snyder R. and Duce P. 2004. Chilling and forcing model to predict
bud - burst of crop and forest species. Agric. For. Meteorol. 126(1-2): 1-13.
doi:10.1016/j.agrformet.2004.03.002
}
