% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilling_units_Harrington.R
\name{chilling_units_Harrington}
\alias{chilling_units_Harrington}
\title{Chilling units (Harrington, 2010)}
\usage{
chilling_units_Harrington(HourTemp, summ = TRUE)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
Function to calculate winter chill for deciduos trees according to the Chilling units model proposed by
Harrington (2010). This function is compatible with some 'chillR' functions (i.e tempResponse) to
estimate metrics from hourly temperature records.
}
\examples{
library(chillR) 

#Example 1
data <- stack_hourly_temps(KA_weather, latitude = 50.62)
chilling_units_Harrington(data$hourtemps$Temp, summ = TRUE)

#Example 2
tempResponse(data, Start_JDay = 345, End_JDay = 58,
models = list(Chill_Harrington = chilling_units_Harrington))
}
\references{
Harrington C., Gould P. and St.Clair J. (2010). Modeling the effects of winter environment on
dormancy release of Douglas-fir. For. Ecol. Manage. 259(4): 798-808
}
