% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_JDay.R
\name{date_to_JDay}
\alias{date_to_JDay}
\title{Transform dates into JDay}
\usage{
date_to_JDay(date, format)
}
\arguments{
\item{date}{Character string or date class elements. This is the date for which the function must return the JDay. The function also
accepts a vector of \code{length > 1}}

\item{format}{Character string. This is the format in which the date(s) are introduced when calling the
function. See \code{\link[base:as.Date]{base::as.Date}}}
}
\description{
In agricultural assessments, the begining and end for a period of interest (e.g. chill accumulation) 
are often expressed as Julian days (JDay) or Day Of the Year (DOY). This function transform dates into 
JDays without the need of looking at external resources for the information
}
\examples{

date_to_JDay(date = c("15/10/2020", "15/10/2019", "01/08/2003"), 
             format = "\%d/\%m/\%Y")

}
