% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dormancy.R
\docType{package}
\name{dormancyR-package}
\alias{dormancyR-package}
\title{dormancyR: a package to compute chill metrics from organized weather dataframes}
\description{
The dormancyR package (inspired in \code{\link{chillR}}) aims to provide alternatives to common chill models
used in horticulture to compute chill metrics in deciduous fruit tree orchards as well as functions to handle
weather data.
}
\author{
Eduardo Fernandez
}
