% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_chill.R
\name{exponential_chill_Tmax}
\alias{exponential_chill_Tmax}
\title{Exponential chilling model (Legave et. al. 2008, 2013)}
\usage{
exponential_chill_Tmax(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing colums "Tmax" and "Tmin". These values must correspond to daily
records.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
This function computes the chill in an exponential way proposed by Legave et. al. (2008) and Legave et. al.
(2013). This model, which uses Tmax as input an a threshold of 15 celsius degree, was selected as one of the
"bests" over several models proposed by the same authors.
}
\references{
Legave J., Farrera I., Almeras T. and Calleja M. 2008. Selecting models of apple flowering time
and understanding how global warming has had and impact on this trait. J. Horticult. Sci. Biotechnol. 83(1):
76 - 84
Legave J., Blanke M., Christen D., Giovannini D, Mathieu V. and Oger R. 2013. A comprehensive overview of the
spatial and temporal variability of apple bud dormancy release and blooming phenology in Western Europe.
Int. J. Biometeorol. 57(2): 317 - 331. doi:10.1007/s00484-012-0551-9
}
