% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_10.R
\name{GDH_10}
\alias{GDH_10}
\title{Growing degree hours model for Tb = 10 celsius degree}
\usage{
GDH_10(HourTemp, summ = TRUE)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures}

\item{summ}{Boolean parameter to add the cumulative summ as output}
}
\description{
Growing degree hours model for Tb = 10 celsius degree
}
