% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_5.R
\name{GDH_5}
\alias{GDH_5}
\title{Growing degree hours model for Tb = 5 celsius degree}
\usage{
GDH_5(HourTemp, summ = TRUE)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures}

\item{summ}{Boolean parameter to add the cumulative summ as output}
}
\description{
Growing degree hours model for Tb = 5 celsius degree
}
