% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_CDC.R
\name{handle_CDC}
\alias{handle_CDC}
\title{Handle historic weather records from stations across Germany}
\usage{
handle_CDC(action, variables, latitude, longitude, begin = 19160101,
  end = chillR::Date2YEARMODA(Sys.Date()), number_of_stations = 25,
  complete_list = FALSE)
}
\arguments{
\item{action}{Character parameter to decide on 3 options. Options are "info_stations", "my_data" and
"list_data". "info_stations" returns a dataframe with the information on close weather stations
to a location. "my_data" downloads the records for the closest weather station. "list_data" downloads
the records for several weather stations which are close to the location of interest.}

\item{variables}{Character vector of the variables required. For now, the function can returns the
wind speed (mean - "Wind_speed" and maximum - "Wind_speed_max"); the atmospheric pressure
("ATM_pressure"); the rainfall ("Rainfall"); the precipitation as snow ("Snow"), the minimum 
temperature 5 cm above the ground ("Tmin_5cm"); the air temperature 2 m above ground (minimum - "Tmin",
mean - "Tmean", and maximum - "Tmax"); the relative humidity ("RH"); and the vapour pressure deficit
(VPD).}

\item{latitude}{Numeric parameter. The latitude (in decimal degrees) of the location of interest.}

\item{longitude}{Numeric parameter. The longitude (in decimal degrees) of the location of interest.}

\item{begin}{Numeric parameter. The start date of the period by which the record is required. It has
to be specified in YEARMODA format.}

\item{end}{Numeric parameter. The end date of the period by which the record is required. It has
to be specified in YEARMODA format.}

\item{number_of_stations}{Numeric parameter. The numbers of stations to select from the point of 
interest.}

\item{complete_list}{Boolean parameter. If "my_data" option has been used this allows to skip the first
weather station when using the "list_data" option. It avoids include a repetead dataframe for the
first weather station.}
}
\description{
It allows to: 1) obtain information of the weather stations from Germany; 2) get weather data for a
place of interest given the coordinates; and 3) get weather data from several weather stations
close to a specific location given its coordinates. The functions take data from the "Climatic Data
Center (CDC - https://cdc.dwd.de/portal/) of Germany.
}
\details{
If "info_stations" is used, the function returns a dataframe (9 columns x number_of_stations) containing
information such as the name, latitude, longitude, begin, end and distance of the weather stations.
If "my_data" is chosen, it downloads the weather data from the CDC website. The output is
a dataframe (in chillR format) containing daily records from the closest
weather station. If "list_data" option is used, the function returns a list of dataframes as that
described above. The length of the list is equal to the number of stations or to the number_of_stations
minus 1 if complete_list = FALSE.
}
\examples{
    
handle_CDC(action = "info_stations", variables = c("Tmin", "Tmax", "Tmean"), latitude = 53.5373,
           longitude = 9.6397, begin = 20000101, end = 20101231, number_of_stations = 25,
           complete_list = FALSE)
           
}
