% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_chile_cr2_prec.R
\name{handle_chile_cr2_prec}
\alias{handle_chile_cr2_prec}
\title{Get historical records of daily precipitation for Chile}
\usage{
handle_chile_cr2_prec(
  action,
  begin = 19500101,
  end = 20180309,
  latitude,
  longitude,
  number_of_stations = 25,
  path_zip_prec,
  stations_df = NULL
)
}
\arguments{
\item{action}{Character string input that defines the kind of data required. There are three options
for this parameter. \emph{"info_stations"} provides a dataframe with information about a given number
of weather stations (set in \code{number_of_stations}) located close to the location
established with \code{latitude} and \code{longitude} parameters. \emph{"list_data"} provides a list of
dataframes containing daily precipitation records from each weather station
obtained with the \emph{"info_stations"} mode within the period established in the call of the function.
Finally, \emph{"my_data"} provides the first dataframe of the previous list and represent the
data of the closest weather station to the location established}

\item{begin}{Numeric parameter in YEARMODA format. This parameter represents the
initial date of the period of interest. If it is not provided, the default is set to
19500101 which correspond to the oldest date possible to use}

\item{end}{Numeric parameter in YEARMODA format. This parameter represents the
final date of the period of interest. If it is not provided, the default is set to
20180309 which corresponds to the earliest possible date that can be used for the assessment}

\item{latitude}{Numerical input. Latitude of the site of interest in decimal format}

\item{longitude}{Numerical input. Longitude of the site of interest in decimal format}

\item{number_of_stations}{Numerical input. Number of stations used as closest weather stations to
the site of interest. Default option is 25}

\item{path_zip_prec}{Character string input. Location and name of the zip file downloaded and
containing the records for precipitation. This input must include the name and extension of the file}

\item{stations_df}{A dataframe containing the list of stations for which this function will retrieve
precipitation data. It is important that this dataframe be produced by this function under the action 
\emph{"info_stations"}. The default is set to \code{NULL}}
}
\description{
This function allows to obtain information about weather stations located in Chile as well as
a data frame containing historical records of daily precipitation from those weather
stations. This function works with data previously downloaded from the website of the
\href{http://www.cr2.cl/}{Center for Climate and Resilience Research (CR)2} sponsored by the University of Chile.
The function works for
downloaded data in ".zip" format. Data can be obtained from the following
\href{http://www.cr2.cl/download/cr2_tasmindaily_2018_ghcn-zip/?wpdmdl=15125}{link}.
}
\examples{
##Getting the location of zip files

#path_zip_prec <- "[Your folder]/cr2_prDaily_2018_ghcn.zip"

##Call of the function

#handle_chile_cr2_prec(action = "my_data", begin = 20000101, end = 20101030,
#                latitude = -32.8958, longitude = -71.2092,
#                number_of_stations = 25,
#                path_zip_prec = path_zip_prec, stations_df = NULL)


}
