% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_CKA.R
\name{handle_CKA}
\alias{handle_CKA}
\title{Get weather data from Campus Klein Altendorf in chillR format}
\usage{
handle_CKA(folder_path, vars = c("Temp"), time_step = "hourly", check_data = T)
}
\arguments{
\item{folder_path}{Chacter string. This is the complete directory name where the files are stored}

\item{vars}{Character string. Variables of interest returned by the function. These are:
"Wind_speed (m/s)", "Wind_direction (degrees)", "Temp (Celsius)", "Tmean (Celsius)", "Tmax (Celsius)",
"Tmin (Celsius)", "Humidity (%)", "Above_Ground_Temp (Celsius)", "Soil_Temp (Celsius)", "Precipitation (mm)",
"Radiation (W/m2)" and "PAR (umol/m2/s)". Default is set to temperature ("Temp")}

\item{time_step}{Character string. This is related to the frequency in which the data was recorded. Normally,
CKA records weather parameters each 10 min. However, this function only returns values for hourly and 
daily time steps. Default is set to "hourly"}

\item{check_data}{Boolean parameter to define wether the data should be quality checked or not. This means to
look for missing hours or days and fill them through linear interpolation. This only works for temperature}
}
\description{
This function allows to handle weather data recorded by weather stations located in Campus Klein Altendorf
in Rhineland, Germany. Normally, the access to the data is restricted to CKA stuff but, any person could ask
for it (https://www.cka.uni-bonn.de/forschung/wetterdaten/aktuelle-wetterdaten). It is important to note that
CKA has the data organized into daily .csv files for the recent period as well as in .zip files for each year.
This function only works for data downloaded and extracted into one common folder. This means, all the files 
need to be in the same folder, independent of the year. The function returns the most important variables for
agricultural production. Among these, temperature, rainfall, wind speed, PAR, among others.
}
\details{
It is important to note that the data has to be downloaded by hand from the website or requested to the stuff
working at CKA. The output is a data frame useful for further analysis with chillR or this same package.
The interpolation is made by using some functions contained in the chillR package. Tmin, Tmean and Tmax
only work for "daily" time step and Temp only does for "hourly" time step.
}
\examples{

# As each user has different path for the folder this example is not running until "#"
# is removed

# path <- "C:/Users/...../...."

# handle_CKA(folder_path = path, vars = c("Tmin", "Tmean", "Tmax"),
# time_step = "daily", check_data = T)

}
