% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_gh_hum_co2.R
\name{handle_gh_hum_co2}
\alias{handle_gh_hum_co2}
\title{Get humidity and concentration of \eqn{CO2} from the heated greenhouse at CKA}
\usage{
handle_gh_hum_co2(path_data, time_step = "hourly")
}
\arguments{
\item{path_data}{Character string. This is the directory in which the file has been stored. It should include
the complete name of the file in ".xls" format.}

\item{time_step}{Character string. Time step in which the data must be returned. It has two options for the
frequency of records. \emph{"hourly"} (the default) and \emph{"daily"}}
}
\description{
This function returns information about relative humidity and concentration of \eqn{CO2} inside the heated
greenhouse located at Campus Klein Altendorf of the University of Bonn. It has two options for time step
(\emph{hourly} and \emph{daily}) basis.
}
\examples{

# As each user has different path for the folder this example is not running until "#"
# is removed

# path <- "C:/Users/...../...."

# handle_gh_hum_co2(path_data = path, time_step = "daily")

}
