% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_GH_Temp.R
\name{handle_GH_Temp}
\alias{handle_GH_Temp}
\title{Get weather data from the greenhouses at CKA in chillR format}
\usage{
handle_GH_Temp(path_data, time_step = "hourly")
}
\arguments{
\item{path_data}{Character string. This is the directory in which the data is stored. It is important to note
that the file must be a ".xls"}

\item{time_step}{Character string. This is the time step in which the data must be obtained. It has two options;
"hourly" (the default) and "daily"}
}
\description{
This function is to handle weather data obtained from the greenhouses located in Campus Klein Altendorf.
For now, this only works for temperatures. Input files must be stored as ".xls" format.
}
\note{
In the "daily" mode, this function returns the variables Tmin, Tmean and Tmax. In contrast, under the "hourly"
mode, it returns the mean temperature for each hour as "Temp".
}
\examples{
# As each user has different path for the folder this example is not running until "#"
# is removed

# path <- "C:/Users/...../...."

# handle_GH_data(path_data = path, time_step = "daily")

}
