% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tinytag.R
\name{handle_tinytag}
\alias{handle_tinytag}
\title{Get weather data from Tinytag sensor in chillR format}
\usage{
handle_tinytag(path_data, vars = c("Temp", "Humidity"),
  time_step = "hourly", check_data = T, latitude = NULL,
  sep = NULL)
}
\arguments{
\item{path_data}{Character string. This is the directory in which the data is stored. It is important to note
that the file must be in ".csv" (using sep = ",") or ".xlsx" format}

\item{vars}{Character string. Climate related variables that will be returned by the function. It has two
options for temperature according to the time_step parameter. Tmean, Tmin and Tmax only work for daily
observations. Temp works for hourly observations. The Humidity is also possible to obtain. Default is set to
"Temp" and "Humidity"}

\item{time_step}{Character string. This is the time step in which the data must be obtained. Default is set to
"hourly"}

\item{check_data}{Boolean parameter to decide wether the data should be quality checked or not. If so, it uses
functions contained in chillR to fill the gaps. Default is TRUE}

\item{latitude}{Numeric vector. If the check_data parameter is used, this is to fill the gaps by linear
interpolation according to the latitude of the place. Default is NULL}

\item{sep}{Sometimes the file is saved in ".csv" format but the character for separating the rows is not ","
(comma). With sep, this can be set to a different separator character only if the user knows it}
}
\description{
This function returns weather data (temperature and humidity) from tinytag sensors (model TGP-4500). It also
contains options to check for missing observations and fill them through linear interpolation.
}
\details{
This function has an option to fill the gaps through linear interpolation. However, this step is ONLY
recommended if the sensor was placed in normal environmental conditions. Otherwise, the values used to fill
the gaps are not representatives for the conditions.
}
\examples{

# As each user has different path for the folder this example is not running until "#"
# is removed

# path <- "C:/Users/...../...."

# handle_tinytag(path_data = path, vars = c("Tmin", "Tmean", "Tmax"),
# time_step = "daily", check_data = T, latitude = 50.62, sep = NULL)

}
