% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_chill.R
\name{rate_of_chill}
\alias{rate_of_chill}
\title{Rate of chilling according to Chmielewski's model}
\usage{
rate_of_chill(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing columns \emph{"Tmax"} and \emph{"Tmin"}. These values must correspond to daily
records}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour}
}
\description{
Function to calculate winter chill for deciduous fruit trees according to the model (sequential model) proposed by
Chmielewski \emph{et al.} (2001) in Germany.
}
\examples{
library(chillR)

#Example 1

rate_of_chill(KA_weather, summ = TRUE)

#Example 2

tempResponse_daily(KA_weather, Start_JDay = 345, End_JDay = 58,
                   models = list(Rate_of_Chill = rate_of_chill))

}
\references{
Chmielewski F., Blumel K., Henniges Y., Blanke M., Weber R. and Zoth M. (2011).
Phenological models for the beginning of apple blossom in Germany. Meteorol. Z. 20(5): 487-496.
\href{https://www.ingentaconnect.com/content/schweiz/mz/2011/00000020/00000005/art00004}{doi:10.1127/0941-2948/2011/0258}
}
