% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_chill_Chmielewski.R
\name{rate_of_chill_Chmielewski}
\alias{rate_of_chill_Chmielewski}
\title{Rate of chilling according to Chmielewski's model}
\usage{
rate_of_chill_Chmielewski(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing colums "Tmax" and "Tmin". These values must correspond to daily
records}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour}
}
\description{
Function to calculate winter chill for deciduos fruit trees according to the model (sequential model) proposed by
Chmielewski et al. (2001) in Germany.
}
\examples{
library(chillR)

#Example 1
rate_of_chill_Chmielewski(KA_weather, summ = TRUE)

#Example 2
tempResponse_daily(KA_weather, Start_JDay = 345, End_JDay = 58,
                   models = list(Rate_of_Chill = rate_of_chill_Chmielewski))

}
\references{
Chmielewski F., Blumel K., Henniges Y., Blanke M., Weber R. and Zoth M. (2011).
Phenological models for the beginning of apple blossom in Germany. Meteorol. Z. 20(5): 487-496.
doi:10.1127/0941-2948/2011/0258
}
