% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempResponse_list_daily.R
\name{tempResponse_list_daily}
\alias{tempResponse_list_daily}
\title{Temperature response for data in lists}
\usage{
tempResponse_list_daily(temperature_list, Start_JDay = 1,
  End_JDay = 366, models = list(Rate_of_Chill =
  rate_of_chill_Chmielewski, Chill_Days = chill_days, Exponential_Chill =
  exponential_chill_Tmax, Triangula_Chill_Hann =
  triangular_chill_daily_Hann, Triangular_Chill_Lega =
  triangular_chill_daily_Tmean))
}
\arguments{
\item{temperature_list}{List of temperatures as those produced by temperature_generation from chillR.}

\item{Start_JDay}{Numeric input representing the initial date for computing the metrics in Julian Day format
This is the number of the day within the year.}

\item{End_JDay}{Numeric input representing the end date for computing the metrics in Julian Day format
This is the number of the day within the year.}

\item{models}{List of functions to compute the metrics. Default is settled to those models used in
tempResponse_daily function. This parameter also allows the use of a model specified by the user.}
}
\description{
This function allows to compute metrics from models which use as input daily temperature istead hourly data.
It is an extension of the tempResponse_daily function.
}
