% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_chill_1.R
\name{triangular_chill_1}
\alias{triangular_chill_1}
\title{Triangular chill function (daily extreme temps as input)}
\usage{
triangular_chill_1(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing columns \emph{"Tmax"} and \emph{"Tmin"}. These values must correspond to daily
records}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour}
}
\description{
Function to calculate winter chill for deciduous trees according to the Triangular chill function proposed by
Hanninen (1990) and Krammer (1994).
}
\examples{
library(chillR)

tempResponse_daily(KA_weather, Start_JDay = 345, End_JDay = 58,
 models = list(Triangular_Chill = triangular_chill_1))
 
}
\references{
Hanninen H. 1990. Modelling bud dormancy release in trees from cool and temperate regions.
Acta For. Fenn. 213: 1-47 doi:10.14214/aff.7660
Kramer K. 1994. Selecting a model to predict the onset of growth of \emph{Fagus sylvatica}.
J. Appl. Ecol. 31(1): 172-181
}
