% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_chill_Legave.R
\name{triangular_chill_Legave}
\alias{triangular_chill_Legave}
\title{Triangular chilling model (Legave et. al. 2008, 2013)}
\usage{
triangular_chill_Legave(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing colums "Tmax" and "Tmin". These values must correspond to daily
records.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
This function computes the chill according to the triangular function proposed by Legave et. al. (2008) and
Legave et. al. (2013). This model, which uses Tmean as input, was selected as one of the "bests" over several
models proposed by the same authors.
}
\examples{
library(chillR)

tempResponse_daily(data, Start_JDay = 345, End_JDay = 58, models = list(Triangular_Chill_Legave =
                                                                        triangular_chill_Legave))
}
\references{
Legave J., Farrera I., Almeras T. and Calleja M. 2008. Selecting models of apple flowering time
and understanding how global warming has had and impact on this trait. J. Horticult. Sci. Biotechnol. 83(1):
76 - 84
Legave J., Blanke M., Christen D., Giovannini D, Mathieu V. and Oger R. 2013. A comprehensive overview of the
spatial and temporal variability of apple bud dormancy release and blooming phenology in Western Europe.
Int. J. Biometeorol. 57(2): 317 - 331. doi:10.1007/s00484-012-0551-9
}
