% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet_data.R
\docType{data}
\name{agromet_weather_data}
\alias{agromet_weather_data}
\title{Example of Chilean Agromet weather data}
\format{
A data frame with 2961 observations and 11 variables
}
\source{
\url{https://agrometeorologia.cl/}
}
\usage{
data(agromet_weather_data)
}
\description{
Data downloaded from the Chilean Agromet network to be used as example in the
\code{\link{handle_agromet_chile}} function. This data contain hourly records from 01-05-2019 to 31-08-2019.
The datafra has been slightly modified (removed the first 5 rows)
}
\examples{

handle_agromet_chile(agromet_weather_data)

}
\keyword{datasets}
