% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agromet_data.R
\docType{data}
\name{agromet_weather_data}
\alias{agromet_weather_data}
\title{Example of Chilean Agromet weather data}
\format{A data frame with 2208 observations and 12 variables}
\source{
\url{https://www.agromet.cl/datos-historicos}
}
\usage{
data(agromet_weather_data)
}
\description{
Data downloaded from the Chilean Agromet network to be used as example in the
\code{\link{chile_agromet3chillR}} function. This data contain hourly records from 01-10-2018 to 31-12-2018.
In order to avoid encoding issues, the names of the columns were modified
}
\examples{

chile_agromet3chillR(agromet_weather_data)

}
\keyword{datasets}
