% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chile_agromet3chillR.R
\name{chile_agromet3chillR}
\alias{chile_agromet3chillR}
\title{Convert a weather file downloaded from the Chilean Agromet website to \code{\link{chillR}} format}
\usage{
chile_agromet3chillR(downloaded_weather_file)
}
\arguments{
\item{downloaded_weather_file}{full path of a weather file downloaded from the Chilean Agromet website
(\url{https://www.agromet.cl/datos-historicos}) as a csv file (\emph{"Datos/Exportar a csv"} option)}
}
\description{
Convert downloaded weather data into a data frame that makes running other \code{\link{chillR}} functions easy.
}
\examples{
data <- Agromet_Weather_data
chile_agromet3chillR(data)

}
