% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chill_days.R
\name{chill_days}
\alias{chill_days}
\title{Chill days (Cesaraccio et al. 2004)}
\usage{
chill_days(ExtrDailyTemp, summ = TRUE)
}
\arguments{
\item{ExtrDailyTemp}{Dataframe containing columns "Tmax" and "Tmin". These values must correspond to daily
records.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
This function computes the Chill Days according to the model proposed by Cesaraccio et al. (2004) developed
for several tree species including pears, kiwifruit and sweet cherries. In this function chill days are
computed without the negative (-) transformation made in the original paper.
}
\examples{
library(chillR)

#Example 1
chill_days(KA_weather, summ = TRUE)

#Example 2
data <- KA_weather
data[,"Chill_Days"] <- chill_days(data, summ = FALSE)

#Example 3
tempResponse_daily(KA_weather, Start_JDay = 345, End_JDay = 58, models = list(Chill_Days = chill_days),
QControl = T) ##what is this QControl? not in the temperature_daily function that's currently in chillR
}
\references{
Cesaraccio C., Spano D., Snyder R. and Duce P. 2004. Chilling and forcing model to predict
bud - burst of crop and forest species. Agric. For. Meteorol. 126(1-2): 1-13.
doi:10.1016/j.agrformet.2004.03.002
}
