% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frost_risk.R
\name{frost_risk}
\alias{frost_risk}
\title{Compute the number of hours below a given threshold}
\usage{
frost_risk(HourTemp, summ = TRUE, threshold = 0)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}

\item{threshold}{Numeric vector below which it is considered a potential frost event. Default to 0 celsius degree.}
}
\description{
This function computes the numbers of hours below a given threshold. This function allows to attempt the
potential risk of spring frost events for deciduous trees. This function is compatible with some 'chillR'
functions (i.e tempResponse) to estimate metrics from hourly temperature records.
}
\note{
By using this function as model input for chillR::tempResponse, the threshold parameter is settled to
default.
}
\examples{
library(chillR) 

data <- stack_hourly_temps(KA_weather, latitude = 50.62)
frost_risk(data$hourtemps$Temp, summ = T, threshold = -1)
}
