% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_agromet_chile.R
\name{handle_agromet_chile}
\alias{handle_agromet_chile}
\title{Handle weather data downloaded from the Agromet Chilean database}
\usage{
handle_agromet_chile(inputfile, vars = c("Temp", "Humidity"))
}
\arguments{
\item{inputfile}{Character string or dataframe. This is the complete path to the weather file containing the data. It must
include the extension (".xls") of the file. The function also supports a dataframe}

\item{vars}{Character string. Vector of one or more variables that are requested from the original weather file.
Default is set to \emph{"Temp"} and \emph{"Humidity"} since for now, the function only works for hourly records.
Available options depend on
the variables specified by the user while downloading from the website. Possible inputs are
\emph{"Temp"}, \emph{"Humidity"}, \emph{"Prec_accum"}, \emph{"Solar_radiation"},
\emph{"Wind_speed"}, \emph{"Pressure_ATM"}, \emph{"Surface_Temp"}, \emph{"Soil_Temp"},
\emph{"Wind_direction"}}
}
\description{
This function convert a previously downloaded weather file into a dataframe in the \code{\link{chillR}}
format. For now,
the function \strong{only} works with hourly records. The downloading part must be done directly by the
user through accessing the
\href{https://agrometeorologia.cl/}{agromet website} Once there, the user needs to select the weather station,
the variables requested, the time-step and the period of interest. Finally, the data must be downloaded in
Excel or CSV format.
}
\examples{

handle_agromet_chile(inputfile = agromet_weather_data, vars = c("Temp", "Humidity"))

}
