% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_Agromet_Chile.R
\name{handle_Agromet_Chile}
\alias{handle_Agromet_Chile}
\title{Handle weahter data downloaded from the Agromet Chilean database}
\usage{
handle_Agromet_Chile(pathfile, vars = c("Temp"))
}
\arguments{
\item{pathfile}{Character string. This is the complete path to the excel file containing the weather data. It must
include the extension ("xls") of the file.}

\item{vars}{Character string. Vector of one or more variables that are requested from the original weather file.
Default is set to "Temp" since for now, the function only works for hourly records. Available options depend on
the variables specified by the user while downloading. Possible inputs are "Precip_Accum", "Humidity",
"Wind_direction", "Solar_radiation", "Pressure_ATM", "Temp", "Tmax", "Tmin", "Soil_Tmean", "Soil_Tmax",
"Wind_speed_max", "Surface_Tmean", "Surface_Tmax", "Soil_Tmin", "Surface_Tmin", "Wind_speed_mean". See details 
for more information.}
}
\description{
This function convert a previously downloaded weather file into a dataframe in the chillR format. For now,
the downloading part must be done directly by the user through accessing the website
http://agromet.inia.cl/estaciones.php. Once there, the user needs to select the weather station,
the variables requested, the time-step and the period of interest. Finally, the data must be downloaded in
excel format and stored. For now, this function only works for files downloaded in the hourly time-step option.
}
\examples{
# As each user has different path for the folder this example is not running until "#"
# is removed

# path <- "C:/Users/...../...."

# handle_Agromet_Chile(pathfile = path, vars = c("Tmin", "Tmean", "Tmax"))

}
