% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_cdc_germany.R
\name{handle_cdc_germany}
\alias{handle_cdc_germany}
\title{Handle historic weather records from stations across Germany}
\usage{
handle_cdc_germany(
  action,
  variables,
  location = NA,
  begin = 19160101,
  end = chillR::Date2YEARMODA(Sys.Date()),
  number_of_stations = 25,
  add.DATE = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{action}{is a character string to decide on 2 modes of action for the function.\itemize{\item{
\emph{"list_stations"} returns a dataframe with the information on close weather stations
to the location defined by \code{number_of_stations} and \code{location} parameters.}
\item{\emph{"download_weather"} retrieves the records for one or more weather stations defined in the
\code{location} parameter.}}}

\item{variables}{is a character vector representing the variables requested For now, the function can return the
wind speed (mean - \emph{"Wind_speed"} and maximum - \emph{"Wind_speed_max"}); the atmospheric pressure
(\emph{"ATM_pressure"}); the rainfall (\emph{"Rainfall"}); the precipitation as snow (\emph{"Snow"}), the minimum 
temperature 5 cm above the ground (\emph{"Tmin_5cm"}); the air temperature 2 m above ground (minimum - \emph{"Tmin"},
mean - \emph{"Tmean"}, and maximum - \emph{"Tmax"}); the relative humidity (\emph{"RH"}); and the vapor pressure deficit
(\emph{"VPD"}).}

\item{location}{accepts a numeric vector with two elements representing the longitude and latitude of a
given place or a vector of character strings representing the ID of the weather stations of interest.
If \code{action = "list_stations"}, \code{location} requires the coordinates of the place. This vector
can be named or not. Accepted names are: \code{"y"}, \code{"Y"}, \code{"latitude"}, \code{"lat"},
\code{"Latitude"}, \code{"Lat"}, \code{"LATITUDE"}, \code{"LAT"} for latitude and \code{"x"},
\code{"X"}, \code{"longitude"}, \code{"long"}, \code{"Longitude"}, \code{"Long"}, \code{"LONGITUDE"},
\code{"LONG"}. If \code{action = "download_weather"}, \code{location} accepts the ID of the station as
character string.}

\item{begin}{is a numeric parameter representing the start date for the period required. This must be
specified in YEARMODA format.}

\item{end}{is a numeric parameter representing the end date for the period required. This must be
specified in YEARMODA format.}

\item{number_of_stations}{is a numeric parameter defininf the numbers of stations to select from.}

\item{add.DATE}{is a boolean parameter to be passed to \code{\link{make_all_day_table}}}

\item{quiet}{is a boolean parameter to be passed to \code{\link[utils:download.file]{download.file}}}
}
\description{
This function access the \href{https://www.dwd.de/EN/climate_environment/cdc/cdc_node.html}{Climate Data Center - CDC} 
and allows to:\itemize{\item{1) obtain information of the weather stations from Germany.}
\item{2) get weather data for one or more specific places of interest given the ID of the weather station.}}
}
\details{
If \emph{"list_stations"} is used, the function returns a dataframe (9 columns x \code{number_of_stations} rows) containing
information such as the name, latitude, longitude, begin, end and distance of the weather stations.
If \emph{"download_weather"} is chosen, it downloads the weather data from the CDC website. The output is
a dataframe (in \code{\link{chillR}} format) for all stations selected in the location parameter.
}
\examples{

handle_cdc_germany(action = "list_stations", variables = c("Tmin", "Tmax", "Tmean"),
                   location = c(latitude = 53.5373, longitude = 9.6397), begin = 20000101,
                   end = 20101231, number_of_stations = 25)

}
