% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_cdc_germany.R
\name{handle_cdc_germany}
\alias{handle_cdc_germany}
\title{Handle historic weather records from stations across Germany}
\usage{
handle_cdc_germany(
  action,
  variables,
  latitude,
  longitude,
  begin = 19160101,
  end = chillR::Date2YEARMODA(Sys.Date()),
  number_of_stations = 25,
  complete_list = FALSE
)
}
\arguments{
\item{action}{Character parameter to decide on 3 options. Options are: \emph{"info_stations"},
\emph{"my_data"} and
\emph{"list_data"}. \emph{"info_stations"} returns a dataframe with the information on close weather stations
to the location defined by \code{longitude} and \code{latitude} parameters. \emph{"my_data"} downloads the records for the
closest weather station. \emph{"list_data"} downloads the records for several weather stations which are close
to the location of interest}

\item{variables}{Character vector of the variables required. For now, the function can returns the
wind speed (mean - \emph{"Wind_speed"} and maximum - \emph{"Wind_speed_max"}); the atmospheric pressure
(\emph{"ATM_pressure"}); the rainfall (\emph{"Rainfall"}); the precipitation as snow (\emph{"Snow"}), the minimum 
temperature 5 cm above the ground (\emph{"Tmin_5cm"}); the air temperature 2 m above ground (minimum - \emph{"Tmin"},
mean - \emph{"Tmean"}, and maximum - \emph{"Tmax"}); the relative humidity (\emph{"RH"}); and the vapour pressure deficit
(\emph{"VPD"})}

\item{latitude}{Numeric parameter. The latitude (in decimal degrees) of the location of interest}

\item{longitude}{Numeric parameter. The longitude (in decimal degrees) of the location of interest}

\item{begin}{Numeric parameter. The start date of the period by which the record is required. It has
to be specified in YEARMODA format}

\item{end}{Numeric parameter. The end date of the period by which the record is required. It has
to be specified in YEARMODA format}

\item{number_of_stations}{Numeric parameter. The numbers of stations to select from the point of 
interest}

\item{complete_list}{Boolean parameter. If \emph{"my_data"} option has been used this allows to skip the first
weather station when using the \emph{"list_data"} option. It avoids include a repeated dataframe for the
first weather station}
}
\description{
This function allows to: 1) obtain information of the weather stations from Germany; 2) get weather data for a
place of interest given the coordinates; and 3) get weather data from several weather stations
close to a specific location given its coordinates. The function takes data from the 
\href{https://www.dwd.de/EN/climate_environment/cdc/cdc_node.html}{Climate Data Center - CDC}
}
\details{
If \emph{"info_stations"} is used, the function returns a dataframe (9 columns x \code{number_of_stations} rows) containing
information such as the name, latitude, longitude, begin, end and distance of the weather stations.
If \emph{"my_data"} is chosen, it downloads the weather data from the CDC website. The output is
a dataframe (in \code{\link{chillR}} format) containing daily records from the closest
weather station. If \emph{"list_data"} option is used, the function returns a list of dataframes as that one
described above. The length of the list is equal to the \code{number_of_stations} or to the \code{number_of_stations}
minus 1 if \code{complete_list = FALSE}.
}
\examples{

handle_cdc_germany(action = "info_stations", variables = c("Tmin", "Tmax", "Tmean"),
                   latitude = 53.5373, longitude = 9.6397, begin = 20000101,
                   end = 20101231, number_of_stations = 25, complete_list = FALSE)

}
