% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_greenhouse.R
\name{handle_greenhouse}
\alias{handle_greenhouse}
\title{Get weather data from the greenhouses at CKA in \code{\link{chillR}} format}
\usage{
handle_greenhouse(
  path_data,
  time_step = "hourly",
  vars = c("Temp", "Humidity", "CO2")
)
}
\arguments{
\item{path_data}{Character string. This is the directory in which the data is stored. It is important to note
that the file must have either \strong{'.xls'} or \strong{'.xlsx'} extension}

\item{time_step}{Character string. This is the time step in which the data must be retrieved. It has two options;
\emph{'hourly'} (the default) and \emph{'daily'}}

\item{vars}{Character string. Variables that must be retrieved by the function. In general, the options are
\emph{'Temperatures'}, \emph{'Humidity'}, and \emph{'CO2 concentration'}. Regarding \emph{'Temperatures'}, 
the options differed depending on the \code{time_step} parameter. For \code{time_step = 'daily'}, the options
are \emph{'Tmean'}, \emph{'Tmin'}, and \emph{'Tmax'}}
}
\description{
This function is to handle weather data obtained from the greenhouses located in Campus Klein Altendorf.
For now, this works for input files stored as \strong{'.xls'} or \strong{'.xlsx'} format
}
\note{
In the \emph{'daily'} mode, this function can returns the variables \emph{'Tmin'}, \emph{'Tmean'} and \emph{'Tmax'}. In contrast, under the \emph{'hourly'}
mode, it can returns the mean temperature for each hour as \emph{'Temp'}
}
\examples{
# As each user has different path for the folder this example is not running until "#"
# is removed

# path <- "C:/Users/...../...."

# handle_greenhouse(path_data = path, time_step = "daily", vars = c("Tmin", "Tmax"))

}
