% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDay_to_date.R
\name{JDay_to_date}
\alias{JDay_to_date}
\title{Transform JDay to date}
\usage{
JDay_to_date(JDay, year, date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{JDay}{Numeric vector. It can be a vector of \code{length = 1} containing integers between 1 and 366}

\item{year}{Numeric input. This is to define the year for which the JDay is transformed to date}

\item{date_format}{Character vector representing the format in which the date should be retrieved. For
more options see \code{\link[base:as.Date]{as.Date}}}
}
\description{
In agricultural assessments, the begining and end for a period of interest (e.g. chill accumulation) 
are often expressed as Julian days (JDay) or Day Of the Year (DOY). However, this makes difficult the
interpretation of the results since most people are used to see dates. This function transform JDays into 
dates
}
\examples{
JDay_to_date(JDay = 67 : 69, year = 2019, date_format = "\%d.\%m.\%Y")

}
