% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_chill_model.R
\name{low_chill_model}
\alias{low_chill_model}
\title{Low-chill model}
\usage{
low_chill_model(HourTemp, summ = TRUE)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour}
}
\description{
Function to compute winter chill for deciduous trees according to the Low-chill model proposed by
Gilreath and Buchanan (1981). This function is compatible with some \code{\link{chillR}} functions
(i.e \code{\link[chillR:tempResponse]{tempResponse}}) to
estimate metrics from hourly temperature records.
}
\examples{
library(chillR) 

#Example 1

data <- stack_hourly_temps(KA_weather, latitude = 50.62)
low_chill_model(data$hourtemps$Temp, summ = TRUE)

#Example 2

tempResponse(data, Start_JDay = 345, End_JDay = 58,
models = list(Low_Chill = low_chill_model))

}
\references{
Gilreath P. and Buchanan D. 1981. Rest prediction model for low-chilling sungold nectarine. J. Am.
Soc. Hort. Sci. 106(4): 426 - 429
}
