% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/north_carolina_model.R
\name{north_carolina_model}
\alias{north_carolina_model}
\title{North Carolina model}
\usage{
north_carolina_model(HourTemp, summ = TRUE)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
Function to calculate winter chill for deciduos trees according to the North Carolina model proposed by
Shaltout and Unrath (1983). This function is compatible with some 'chillR' functions (i.e tempResponse) to
estimate metrics from hourly temperature records.
}
\examples{
#Example 1
library(chillR)
data <- stack_hourly_temps(KA_weather, latitude = 50.62)
north_carolina_model(data$hourtemps$Temp, summ = TRUE)

#Example 2
tempResponse(data, Start_JDay = 345, End_JDay = 58,
             models = list(North_Carolina_Units = north_carolina_model))

}
\references{
Shaltout A. and Unrath C. 1983. Rest completion prediction model for starkrimsom delicious apples.
J. Am. Soc. Hort. Sci. 108(6): 957-961
}
