% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_utah_model.R
\name{positive_utah_model}
\alias{positive_utah_model}
\title{Positive Utah model}
\usage{
positive_utah_model(HourTemp, summ = TRUE)
}
\arguments{
\item{HourTemp}{Vector of hourly temperatures.}

\item{summ}{Boolean parameter indicating whether the computed metric should be provided as cumulative values
over the period or as the actual accumulation for each hour.}
}
\description{
Function to calculate winter chill for deciduos trees according to the Positive Utah model proposed by
Linsley-Noakes and Allan (1994). This function is compatible with some 'chillR' functions (i.e tempResponse)
to estimate metrics from hourly temperature records.
}
\examples{
#Example 1
data <- stack_hourly_temps(KA_weather, latitude = 50.62)
positive_utah_model(data$hourtemps$Temp, summ = T)

#Example 2
tempResponse(data, Start_JDay = 345, End_JDay = 58,
             models = list(Positive_Units = positive_utah_model))
}
\references{
Linsley-Noakes G. and Allan P. 1994. Comparison of two models for the prediction of rest
competion in peaches. Sci. Hortic. 59(2): 107-113
}
