% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempResponse_daily.R
\name{tempResponse_daily}
\alias{tempResponse_daily}
\title{Summarize chill responses from daily chill functions}
\usage{
tempResponse_daily(data, Start_JDay = 1, End_JDay = 366,
  models = list(Rate_of_Chill = rate_of_chill_Chmielewski, Chill_Days =
  chill_days, Exponential_Chill = exponential_chill_Tmax,
  Triangula_Chill_Haninnen = triangular_chill_Hanninen,
  Triangular_Chill_Legave = triangular_chill_Legave))
}
\arguments{
\item{data}{Could be a list of two or more elements. This list should contain a dataframe that could be
called "weather" if this is obtained from the fix_weather function in chillR or "data" is the list is
obtained from the chile_weather function in "eduaRdo" package. A third option is just a dataframe.
In any case, dataframe should contains colums "Year", "Month", "Day", "JDay", "Tmin" and "Tmax".}

\item{Start_JDay}{The start date for computing the outputs. This is the number of the day within the year.}

\item{End_JDay}{The end date for computing the outputs. This is the number of the day within the year.}

\item{models}{A list of chill functions to compute the metric for the period of interest. Each model
should be named. Default provides five models (See datails).}
}
\description{
This function produces a summarized dataframe of chill responses computed by several chill models which
estimate daily chill instead of the metric for each hour. Five models are predetermined, but this function
also allows to include models predifined by the user. The only request is that such model function must
uses daily mean, max or min temperatures as input.
}
\examples{
library(chillR)

tempResponse_daily(KA_weather, Start_JDay = 335, End_JDay = 58)
}
