% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempResponse_list_daily.R
\name{tempResponse_list_daily}
\alias{tempResponse_list_daily}
\title{Daily temperature responses for data in lists}
\usage{
tempResponse_list_daily(
  temperature_list,
  Start_JDay = 1,
  End_JDay = 366,
  models = list(Rate_of_Chill = rate_of_chill, Chill_Days = chill_days, Exponential_Chill
    = exponential_chill, Triangula_Chill_Haninnen = triangular_chill_1,
    Triangular_Chill_Legave = triangular_chill_1),
  misstolerance = 20
)
}
\arguments{
\item{temperature_list}{List of temperatures as those produced by \code{\link[chillR:temperature_generation]{chillR::temperature_generation}}}

\item{Start_JDay}{Numeric input representing the initial date for computing the metrics in Julian Day format
This is the number of the day within the year}

\item{End_JDay}{Numeric input representing the end date for computing the metrics in Julian Day format
This is the number of the day within the year}

\item{models}{List of functions to compute the metrics. Default is settled to those models used in
\code{\link{tempResponse_daily}} function. This parameter also allows the use of a model specified by the user}

\item{misstolerance}{is the threshold defined by the user to include seasons having days with missing data.
Default is set to 20}
}
\description{
This function allows to compute metrics from models which use daily temperature as inputs, rather than hourly data.
It is an extension of the \code{\link{tempResponse_daily}} function.
}
\examples{
library(chillR)

tempResponse_list_daily(KA_weather, Start_JDay = 345, End_JDay = 58)

}
